Initiated perf event
This perf event will just send a value of 1 to the user space, to notify
that the map repairSymbolBuffer contains the repair symbol as payload
(for the XOR). This commit:
1) Creates the perf event buffer in the BPF program and send a
notification when the repair symbols can be send in user space
2) Creates and handles the perf event buffer in the user space.
"Handles" in the sense that we enter the correct function when an event
is received.

TODO: get the coded data and send as payload of a new IPv6 SRH packet