Change perf event + TODOs
Change the perf event to submit directly the repairSymbols and not just
a notificaiton. This change is to be safe in case a source packet comes
before the repair symbol is handled in the user space. In that case, the
repairSymbol stored in repairSymbolBuffer could be erased before beeing
sent => bad behavior.