The decoder seems to send correct packets
According to Wireshark. Still to do: manually compute the next address
from current address. Also change the port because for now it thinks it
is a DNS request. To test if the decoder correctly sends packets, I
change the BPF_OK by BPF_DROP at the very end of the decoder.bpf.c to
ensure that the only related packets are the recovered ones