Add files RLC + comment on previous commit
Add files for "clean" RLC based on the github of Francois Michel
(UCLouvain). Also implemented the sender RLC in the user space because
too much iterations for the kernel space + I think it will remove some
overhead from the kernel space.

Tested in the sense that the repair symbols have the correct TLV, but I
do not know yet if the coded payload is correct, must see with the
decoder