Recover correct packet size
Only tested with packets of the same size but should work correctly for
all examples. I encode and decode using the same buffers as before, just
extended to include the length at the end of the buffers. The coded
payload length is carried by the tlvRepair__convo_t struct, field
coded_payload_len